✨ feat: complete the enterprise information, periodTime, notification

+++ yunnan-enterprise/src/main/java/com/yunnanenterprise/service/report/PeriodUtils.java
@@ -5,7 +5,8 @@ package com.yunnanenterprise.service.report;
  * 选择 yyyymm 是因为 period 粒度是“月”，int 存储简单且唯一。
  */
 public final class PeriodUtils {
-    private PeriodUtils() {}
+    private PeriodUtils() {
+    }
 
     public static Integer toPeriodId(String yyyyMm) {
         if (yyyyMm == null || yyyyMm.length() != 7 || yyyyMm.charAt(4) != '-') {
@@ -15,10 +16,12 @@ public final class PeriodUtils {
         return Integer.parseInt(compact);
     }
 
-    public static String fromPeriodId(Integer periodId) {
-        if (periodId == null) return null;
+    public static String fromPeriodId(Long periodId) {
+        if (periodId == null)
+            return null;
         String s = String.valueOf(periodId);
-        if (s.length() != 6) return s;
+        if (s.length() != 6)
+            return s;
         return s.substring(0, 4) + "-" + s.substring(4, 6);
     }
 }
